\name{myPackageSkeleton-package}
\alias{myPackageSkeleton-package}
\alias{myPackageSkeleton}
\docType{package}
\title{
\packageTitle{myPackageSkeleton}
}
\description{
\packageDescription{myPackageSkeleton}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{myPackageSkeleton}
\packageIndices{myPackageSkeleton}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{myPackageSkeleton}

Maintainer: \packageMaintainer{myPackageSkeleton}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
